ai_service/requirements.txt
fastapi
uvicorn
pydantic
numpy
huggingface-hub
python-dotenv
